fix: do not call admission webhook update on CRD deletion
see https://gravitee.atlassian.net/browse/GKO-1208

(cherry picked from commit a2bbd0edf4a1e3b4481c7aa0d5119d9891b0be32)

+++ internal/admission/api/v4/validate.go
@@ -133,6 +133,10 @@ func validateUpdate(
 		return errs
 	}
 
+	if newApi.IsBeingDeleted() {
+		return errs
+	}
+
 	base.DeleteDefinitionConfigMapIfNeeded(ctx, oldApi, newApi)
 
 	errs.Add(validateApiType(oldApi, newApi))
